[core] Cleaned up Transform
Eliminated many redundant methods on Transform. Minimized usage of TransformState::x and TransformState::y. Added convenient constructors for AnimationOptions. When scaling, passing in view’s origin as the anchor no longer anchors the scaling operation at the view’s center. To specify no anchor, use NaN.

Increased precision of worldSize(), now that it’s used much more heavily in transform methods.

+++ src/mbgl/map/transform_state.hpp
@@ -96,7 +96,7 @@ private:
     double latY(double lat) const;
     double zoomScale(double zoom) const;
     double scaleZoom(double scale) const;
-    float worldSize() const;
+    double worldSize() const;
 
     mat4 coordinatePointMatrix(double z) const;
     mat4 getPixelMatrix() const;
